+++ backend/AAiT-backend-group-1/usecases/blogUseCase.go
@@ -1,0 +1,79 @@
+package usecases
+
+import "github.com/RealEskalate/a2sv-g5-project-phase-starter-project/aait-backend-group-1/domain"
+
+
+type blogUseCase struct {
+	blogRepository domain.BlogRepository
+}
+
+func NewBlogUseCase(br domain.BlogRepository) domain.BlogUseCase {
+	return &blogUseCase{
+		blogRepository: br,
+	}
+}
+
+func (bu *blogUseCase) CreateBlog(blog *domain.Blog) error {
+	// Implement the logic for creating a blog
+	return nil
+}
+
+func (bu *blogUseCase) GetBlog(blogID string) (*domain.Blog, error) {
+	// Implement the logic for getting a blog by ID
+	return nil, nil
+}
+
+func (bu *blogUseCase) GetBlogs() ([]*domain.Blog, error) {
+	// Implement the logic for getting all blogs
+	return nil, nil
+}
+
+func (bu *blogUseCase) UpdateBlog(blogID string, blog *domain.Blog) error {
+	// Implement the logic for updating a blog
+	return nil
+}
+
+func (bu *blogUseCase) DeleteBlog(blogID string) error {
+	// Implement the logic for deleting a blog
+	return nil
+}
+
+func (bu *blogUseCase) SearchBlogs(title, author string) ([]*domain.Blog, error) {
+	// Implement the logic for searching blogs by title and author
+	return nil, nil
+}
+
+func (bu *blogUseCase) FilterBlogs(tags []string, dateAfter time.Time, popular bool) ([]*domain.Blog, error) {
+	// Implement the logic for filtering blogs by tags, date, and popularity
+	return nil, nil
+}
+
+func (bu *blogUseCase) LikeBlog(userID, blogID string) error {
+	// Implement the logic for liking a blog
+	return nil
+}
+
+func (bu *blogUseCase) AddComment(blogID string, comment *domain.Comment) error {
+	// Implement the logic for adding a comment to a blog
+	return nil
+}
+
+func (bu *blogUseCase) DeleteComment(blogID, commentID string) error {
+	// Implement the logic for deleting a comment from a blog
+	return nil
+}
+
+func (bu *blogUseCase) EditComment(commentID string, comment *domain.Comment) error {
+	// Implement the logic for editing a comment
+	return nil
+}
+
+func (bu *blogUseCase) Like(blogID, userID string) error {
+	// Implement the logic for liking a blog
+	return nil
+}
+
+func (bu *blogUseCase) DisLike(blogID, userID string) error {
+	// Implement the logic for disliking a blog
+	return nil
+}
